+++ src/main/java/shop/nuribooks/view/admin/coupon/bookcoupon/dto/BookCouponResponse.java
@@ -1,0 +1,4 @@
+package shop.nuribooks.view.admin.coupon.bookcoupon.dto;
+
+public record BookCouponResponse(Long couponId) {
+}
