+++ src/main/java/shop/nuribooks/view/admin/coupon/bookcoupon/service/BookCouponService.java
@@ -1,0 +1,13 @@
+package shop.nuribooks.view.admin.coupon.bookcoupon.service;
+
+import shop.nuribooks.view.admin.coupon.bookcoupon.dto.BookCouponRequest;
+import shop.nuribooks.view.admin.coupon.bookcoupon.dto.BookCouponResponse;
+import shop.nuribooks.view.common.dto.ResponseMessage;
+
+public interface BookCouponService {
+
+	ResponseMessage registerBookCoupon(BookCouponRequest bookCouponRequest);
+
+	BookCouponResponse getBookCoupon(Long id);
+
+}
