+++ src/main/java/shop/nuribooks/view/admin/coupon/categorycoupon/dto/CategoryCouponResponse.java
@@ -1,0 +1,4 @@
+package shop.nuribooks.view.admin.coupon.categorycoupon.dto;
+
+public record CategoryCouponResponse(Long couponId) {
+}
