--- src/main/java/shop/nuribooks/view/admin/coupon/dto/BookCouponResponse.java
@@ -1,4 +1,0 @@
-package shop.nuribooks.view.admin.coupon.dto;
-
-public record BookCouponResponse(Long couponId) {
-}
