--- src/main/java/shop/nuribooks/view/admin/coupon/dto/CategoryCouponResponse.java
@@ -1,4 +1,0 @@
-package shop.nuribooks.view.admin.coupon.dto;
-
-public record CategoryCouponResponse(Long couponId) {
-}
