+++ src/main/java/shop/nuribooks/view/admin/coupon/service/CouponService.java
@@ -3,10 +3,6 @@ package shop.nuribooks.view.admin.coupon.service;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import shop.nuribooks.view.admin.coupon.dto.BookCouponRequest;
-import shop.nuribooks.view.admin.coupon.dto.BookCouponResponse;
-import shop.nuribooks.view.admin.coupon.dto.CategoryCouponRequest;
-import shop.nuribooks.view.admin.coupon.dto.CategoryCouponResponse;
 import shop.nuribooks.view.admin.coupon.dto.CouponRequest;
 import shop.nuribooks.view.admin.coupon.dto.CouponResponse;
 import shop.nuribooks.view.admin.coupon.dto.MemberCouponIssueRequest;
@@ -18,19 +14,11 @@ public interface CouponService {
 
 	Page<CouponResponse> getAllCoupons(Pageable pageable);
 
-	BookCouponResponse getBookCoupon(Long id);
-
-	CategoryCouponResponse getCategoryCoupon(Long id);
-
 	ResponseMessage registerCoupon(CouponRequest couponRequest);
-
-	ResponseMessage registerBookCoupon(BookCouponRequest bookCouponRequest);
 
 	ResponseMessage updateCoupon(Long id, CouponRequest couponRequest);
 
 	ResponseMessage expireCoupon(Long id);
-
-	ResponseMessage registerCategoryCoupon(CategoryCouponRequest couponRequest);
 
 	CouponResponse getCouponById(Long couponId);
