+++ src/main/java/shop/nuribooks/view/book/controller/BookController.java
@@ -21,8 +21,10 @@ import lombok.extern.slf4j.Slf4j;
 import shop.nuribooks.view.admin.category.dto.CategoryRequest;
 import shop.nuribooks.view.admin.category.dto.CategoryResponse;
 import shop.nuribooks.view.admin.category.service.AdminCategoryService;
-import shop.nuribooks.view.admin.coupon.dto.BookCouponResponse;
-import shop.nuribooks.view.admin.coupon.dto.CategoryCouponResponse;
+import shop.nuribooks.view.admin.coupon.bookcoupon.dto.BookCouponResponse;
+import shop.nuribooks.view.admin.coupon.bookcoupon.service.BookCouponService;
+import shop.nuribooks.view.admin.coupon.categorycoupon.dto.CategoryCouponResponse;
+import shop.nuribooks.view.admin.coupon.categorycoupon.service.CategoryCouponService;
 import shop.nuribooks.view.admin.coupon.service.CouponService;
 import shop.nuribooks.view.book.dto.BookContributorsResponse;
 import shop.nuribooks.view.book.dto.BookResponse;
@@ -45,6 +47,8 @@ public class BookController {
 	private final AdminCategoryService adminCategoryService;
 	private final BookLikeService bookLikeService;
 	private final CouponService couponService;
+	private final BookCouponService bookCouponService;
+	private final CategoryCouponService categoryCouponService;
 	private final String RECENT_VIEW_LIST_KEY = "recent_view_list";
 
 	@GetMapping("/view/books")
@@ -81,7 +85,7 @@ public class BookController {
 		model.addAttribute("likeStatus", likeStatus);
 
 		try {
-			BookCouponResponse bookCoupon = couponService.getBookCoupon(bookId);
+			BookCouponResponse bookCoupon = bookCouponService.getBookCoupon(bookId);
 			model.addAttribute("bookCoupon", bookCoupon);
 		} catch (Exception e) {
 			model.addAttribute("bookCoupon", null);
@@ -105,7 +109,7 @@ public class BookController {
 		model.addAttribute("categoryId", categoryId);
 
 		try {
-			CategoryCouponResponse categoryCoupon = couponService.getCategoryCoupon(categoryId);
+			CategoryCouponResponse categoryCoupon = categoryCouponService.getCategoryCoupon(categoryId);
 			model.addAttribute("categoryCoupon", categoryCoupon);
 		} catch (Exception e) {
 			model.addAttribute("categoryCoupon", null);
